test: move api tests into subfolder
