+++ tests/api/broadcast.test.js
@@ -3,8 +3,8 @@ import { createTree, waitForPath } from "../util/trees.js";
 import { treeTeardown } from "../util/context.js";
 import { diff } from "../util/snapshot.js";
 
-import single from "../specimens/single.js";
-import grandchild from "../specimens/grandchild.js";
+import single from "./specimens/single.js";
+import grandchild from "./specimens/grandchild.js";
 
 describe("broadcast", (it) => {
     it.after.each(treeTeardown);
@@ -33,24 +33,14 @@ describe("broadcast", (it) => {
         ]`);
     });
 
-    it.only("should send to child trees", async (context) => {
-        const tree = context.tree = createTree(grandchild, { verbose : true });
-
-        console.log("\n\nAWAITING TREE FOR BEFORE\n\n");
+    it("should send to child trees", async (context) => {
+        const tree = context.tree = createTree(grandchild);
 
         const before = await tree();
 
-        console.log("\n\nGOT TREE FOR BEFORE\n\n");
-
-        console.log("\n\nSENDING NEXT TO TREE\n\n")
-
         tree.builder.broadcast("NEXT");
 
-        console.log("\n\nAWAITING TREE FOR AFTER\n\n");
-
         const after = await waitForPath(tree, "grandchild.two");
-
-        console.log("\n\nGOT TREE FOR AFTER\n\n");
 
         diff(before, after, `
         [
